requirement.txt
paddleocr
langchain
ollama
pillow